Use Debputy to lint and reformat (replacing wrap-and-sort)
Debputy offers more extensive static analysis of debian/ contents than
any other tool. Use it to validate the Debian packaging contents. Also
use it to replace `wrap-and-sort` and the custom debian/patches checker
as Debputy handles all of those checks.

Also remove the ''#!/usr/bin/make -f' check in debian/rules as Lintian
checks for this already.

Rewrite the errors to just emit text, no exit codes, and not to re-run
the tool as it may be slow and unnecessary.